move cast function in utils
